add txnctx (#8520)
1. Add [txnCtx](https://github.com/matrixorigin/docs/blob/main/tech-notes/sql/context.md)
2. Clean code

Approved by: @iamlinjunhong
