+++ pkg/frontend/authenticate.go
@@ -22,6 +22,7 @@ import (
 	"math"
 	"math/bits"
 	"os"
+	"strconv"
 	"strings"
 	"sync/atomic"
 	"time"
@@ -64,9 +65,13 @@ type TenantInfo struct {
 }
 
 func (ti *TenantInfo) String() string {
+	delimiter := ti.delimiter
+	if !strconv.IsPrint(rune(delimiter)) {
+		delimiter = ':'
+	}
 	return fmt.Sprintf("{account %s%c%s%c%s -- %d%c%d%c%d}",
-		ti.Tenant, ti.delimiter, ti.User, ti.delimiter, ti.DefaultRole,
-		ti.TenantID, ti.delimiter, ti.UserID, ti.delimiter, ti.DefaultRoleID)
+		ti.Tenant, delimiter, ti.User, delimiter, ti.DefaultRole,
+		ti.TenantID, delimiter, ti.UserID, delimiter, ti.DefaultRoleID)
 }
 
 func (ti *TenantInfo) GetTenant() string {
@@ -5369,7 +5374,9 @@ func InitSysTenant(ctx context.Context, autoincrcaches defines.AutoIncrCaches) e
 		return err
 	}
 	defer mpool.DeleteMPool(mp)
-	bh := NewBackgroundHandler(ctx, mp, pu, autoincrcaches)
+	//Note: it is special here. The connection ctx here is ctx also.
+	//Actually, it is ok here. the ctx is moServerCtx instead of requestCtx
+	bh := NewBackgroundHandler(ctx, ctx, mp, pu, autoincrcaches)
 	defer bh.Close()
 
 	//USE the mo_catalog
@@ -5909,32 +5916,6 @@ func createTablesInInformationSchemaOfGeneralTenant(ctx context.Context, bh Back
 		}
 	}
 	return err
-}
-
-func checkUserExistsOrNot(ctx context.Context, pu *config.ParameterUnit, tenantName string) (bool, error) {
-	mp, err := mpool.NewMPool("check_user_exists", 0, mpool.NoFixed)
-	if err != nil {
-		return false, err
-	}
-	defer mpool.DeleteMPool(mp)
-
-	sqlForCheckUser, err := getSqlForPasswordOfUser(ctx, tenantName)
-	if err != nil {
-		return false, err
-	}
-
-	// A mock autoIncrCaches
-	aic := defines.AutoIncrCaches{}
-	erArray, err := executeSQLInBackgroundSession(ctx, mp, pu, sqlForCheckUser, aic)
-	if err != nil {
-		return false, err
-	}
-
-	if !execResultArrayHasData(erArray) {
-		return false, nil
-	}
-
-	return true, nil
 }
 
 // InitUser creates new user for the tenant
