+++ pkg/frontend/session_test.go
@@ -67,11 +67,16 @@ func TestTxnHandler_NewTxn(t *testing.T) {
 			CommitOrRollbackTimeout: time.Second,
 		}).AnyTimes()
 
+		pu, err := getParameterUnit("test/system_vars_config.toml", eng, txnClient)
+		convey.So(err, convey.ShouldBeNil)
+
 		txn := InitTxnHandler(eng, txnClient)
 		txn.ses = &Session{
 			requestCtx: ctx,
+			pu:         pu,
+			connectCtx: ctx,
 		}
-		err := txn.NewTxn()
+		err = txn.NewTxn()
 		convey.So(err, convey.ShouldBeNil)
 		err = txn.NewTxn()
 		convey.So(err, convey.ShouldNotBeNil)
@@ -110,11 +115,16 @@ func TestTxnHandler_CommitTxn(t *testing.T) {
 
 		txnClient.EXPECT().New().Return(txnOperator, nil).AnyTimes()
 
+		pu, err := getParameterUnit("test/system_vars_config.toml", eng, txnClient)
+		convey.So(err, convey.ShouldBeNil)
+
 		txn := InitTxnHandler(eng, txnClient)
 		txn.ses = &Session{
 			requestCtx: ctx,
+			pu:         pu,
+			connectCtx: ctx,
 		}
-		err := txn.NewTxn()
+		err = txn.NewTxn()
 		convey.So(err, convey.ShouldBeNil)
 		err = txn.CommitTxn()
 		convey.So(err, convey.ShouldBeNil)
@@ -155,11 +165,16 @@ func TestTxnHandler_RollbackTxn(t *testing.T) {
 
 		txnClient.EXPECT().New().Return(txnOperator, nil).AnyTimes()
 
+		pu, err := getParameterUnit("test/system_vars_config.toml", eng, txnClient)
+		convey.So(err, convey.ShouldBeNil)
+
 		txn := InitTxnHandler(eng, txnClient)
 		txn.ses = &Session{
 			requestCtx: ctx,
+			pu:         pu,
+			connectCtx: ctx,
 		}
-		err := txn.NewTxn()
+		err = txn.NewTxn()
 		convey.So(err, convey.ShouldBeNil)
 		err = txn.RollbackTxn()
 		convey.So(err, convey.ShouldBeNil)
@@ -194,6 +209,7 @@ func TestSession_TxnBegin(t *testing.T) {
 		eng.EXPECT().Commit(gomock.Any(), gomock.Any()).Return(nil).AnyTimes()
 		session := NewSession(proto, nil, config.NewParameterUnit(&config.FrontendParameters{}, eng, txnClient, nil), gSysVars, false)
 		session.SetRequestContext(context.Background())
+		session.SetConnectContext(context.Background())
 		return session
 	}
 	convey.Convey("new session", t, func() {
@@ -519,6 +535,7 @@ func TestSession_TxnCompilerContext(t *testing.T) {
 		proto := NewMysqlClientProtocol(0, ioses, 1024, sv)
 		session := NewSession(proto, nil, pu, gSysVars, false)
 		session.SetRequestContext(context.Background())
+		session.SetConnectContext(context.Background())
 		return session
 	}
 
@@ -598,6 +615,7 @@ func TestSession_GetTempTableStorage(t *testing.T) {
 		proto := NewMysqlClientProtocol(0, ioses, 1024, sv)
 		session := NewSession(proto, nil, pu, gSysVars, false)
 		session.SetRequestContext(context.Background())
+		session.SetConnectContext(context.Background())
 		return session
 	}
